Do not use venv lock for uv venv
Fixes a dead lock. The bazel process which is started will hold the venv
lock, even after the execution of the make target would finish.

I assume the lock is not needed if the whole venv build procedure is
under the control of bazel. Bazel cares about concurrent invocations
which pipenv did not so well.

Change-Id: Iede293b80b8f12420c6588777bbd80c99899b3f7