check_http: address embedding issue
Sadly, we can't safely migrate fields, which don't have an explicit
`address` field. The reason is that `$HOSTADDRESS$` does not include the
brackets `[$HOSTADDRESS$]`, which are necessary to form a valid url, if
`$HOSTADDRESS$` is an IPv6 address.

We can't add the brackets on the fly either, since the check_http might
match on an IPv4 and and IPv6 host, in which case no single consistent
rule exists.

CMK-20616

Change-Id: I3eb88e1169d9bbc8fb9e20a3a7593956ac2fafd9