Fix editing Event Console global settings
ConfigDomainEventConsole can not be instanciated ad-hoc anymore because
it requires an argument which is not available in the context of the
global setting. We now refer to the instance registered to
config_domain_registry.

Had to change the return type of ConfigVariable.domain() to provide an
instance instead of the class to get this fixed.

CMK-20610

Change-Id: Ie4e0b94810d976047bc7189f7b313f46af51d06b