17317 FIX Licensing: Clean up incompatible license information during upgrade
If an already licensed Checkmk site is upgraded to an edition where a license is required (Cloud, Managed), an incompatible license would mean that certain functionality is restricted (e.g. activation of changes) is immediately restricted.
To make the upgrade process easier, the incompatible license is now instead removed, requiring a suitable license to be applied.

To see what license is compatible with which edition, see [werk #17310](https://checkmk.com/werk/17310).

CMK-20427

Change-Id: Ifedd15078963cc9d706c87c761613011399eb634