FIX Support Diagnostics housekeeping
The diagnostics dump job splits collecting data into chunks and creates
a tar for each chunk. The internal housekeeping, on the other hand, only
keeps 5 tar files.
So, when lots of options are selected to be included into the dump, the
tars were deleted before they were collected.

Reference: CMK-20621

The number of tars to keep is now set to 10.

Change-Id: Ifaa1d055d7b2971c29a4a7c66a455fff111bd6bd