gui_e2e: fix and unskip test_add_top_list_dashlet
The amount of rows in the dashlet table depends on the amount of
hosts and services. Since we only know the amount of hosts, the
check is updated to only test that the amount of rows in the table
can be devided by the amount of hosts without a remainder.

Jira ticket: CMK-20662

Change-Id: I47f216a92578eeb26d5f35cdadd9bcc363100618