check-cert: go back to using time::Duration instead of std::Duration
this partially reverts debeaa712ab3819259d26790ce727d8f55ce413b

The deprecation warning were the result of using time::Instant, with
this change we will still use std::Instant but convert the elapsed time
to time::Duration.

The reason for using time::Duration is that std::Duration does not have
the Display trait implemented which is needed in a follow-up change.

CMK-17764

Change-Id: If38648886a7a5f412d2253a8b082093a7dd3b464