No longer create users for automation secrets
When we load all users we scan through the profiledir to collect
information. Previously when we encountered an automation secret we
created an user for that directory.
However creating a user is not so simple (anymore) so when you "created"
a user that way, e.g.:

    mkdir var/check_mk/web/foobar
    echo "foo" > var/check_mk/web/foobar/automation.secret

It showed that user in the UI but it crashed since other info about that
user was expected.

Change-Id: I2afd916e523c3438e154577be979ee67f85e6cb6