Make mknotifyd package more self contained
Some static files were still in the generic enterprise skel directory,
which can now be handled by bazel in the package itself.

Introduce a bazel way to deal with the monolithic skel.permissions file.
This file is needed by OMD to later set file permissions when creating
the sites etc directory and files based on the skel hierarchy. This
piece will be needed once we move more packages towards bazel.

Change-Id: Iba31ceae1e27fccc58b88df35e4c020c27185a60