Fix: mk-job (linux) support versions without perl again
Current Linux Distributions like RHEL 8 + 9 do not install Perl as default.
So with the Update Checkmk 2.3, mk-job fails there since an earlier workaround for aix/ solaris to get a timestamp using perl, was made a default now for as well Linux.

This fix tries to restore the function on GNU Linux again if perl is missing, but without reverting the use of perl as default.
Since GNU date should always have the %s parameter, just BSD or BusyBox can lack the support, that's maybe a reasonable approach.

CMK-20768
closes: #775

Change-Id: Icf9e182511ef7fa0898883710189b2fcf0960e66