testlib.playwright.pom.login: Allow testing CSE edition
There is no need to login for the CSE edition. In future
versions we take care of this via a redisigned login process,
but refactoring all 2.3.0 tests is not feasible at this point.

Therefore this change makes sure the login is skipped for the
CSE edition and testing can continue normally.

Change-Id: Idbe3fe1253cb072a16b945c435911c4068e40fd8