htmllib: render href attributes correctly
When rendering an `href` value the ampersand (&) sign was being wrongly
escaped, leading to broken links.

Given that the application uses the `makeuri_contextless` helper
throughout the app, we can be fairly confident that the urls being
were already url encoded. As a fail safe, we will also check
whether the url passed is valid.

CMK-19495

Change-Id: I4d34dec163c3f0861746b00b1fd19f17844bd2dd