17419 FIX check-httpv2: add options to disable certificates & to connect to physical host
With this werk, we have introduced two new optional parameters.

1. short option: -D,  long option: --disable-cert

You should think very carefully before using this parameter. If
invalid certificates are trusted, any certificate for any
certificate for any site will be trusted for use. This includes
expired certificates. This also introduces significant
vulnerabilities, and should only be used as a last resort.

2. short option: -p, long option: --server

With this option, the behaviour is similar to the --resolve
parameter used with curl. It will resolve to the phsyical
address given.

Change-Id: I025a41e361ad872f43cf27ff62f07d7e9827ec53