FIX Support Diagnostics housekeeping
The diagnostics dump job splits collecting data into chunks and creates
a tar for each chunk. The internal housekeeping, on the other hand, only
keeps 5 tar files.
So, when lots of options are selected to be included into the dump, the
tars were deleted before they were collected.

Reference: CMK-20621

The number of tars to keep is now set to 10.

Change-Id: If96f97a95958824ffcc72deb44b70f9bc287e02c