FormSpecs: Introduce extended versions
MultipleChoiceElementExtended and SingleChoiceElementExtended are being
added to be able to skip the public __post_init__ restriction, which
does not allow non valid Python identifiers.

This fixes a potential crash when loading the notifications quick setup.

When for example contact group or users contained non valid Python
identifiers as names loading the notifications quick setup would crash.

And example for this would be an email address as a user id
(foo.bar@foobar.com)

With this change the CascadingSingleChoiceElementExtended and
MultipleChoiceElementExtended are being used, which lift this
restriction.

CMK-20351

Change-Id: I264e9f8c73ce4c417e300edc25e59b3aed8cd8ee