16124 FIX The alert handler master control has an immediate effect now in the CMC
Due to a regression in Checkmk 2.0, the master control switch for alert
handlers only had an effect after activating changes or reloading/restarting
the CMC.  This fix restores the old immediate behavior. Note that users of
the Nagios core were not affected.

Technically, the Livestatus commands ENABLE_ALERT_HANDLERS and
DISABLE_ALERT_HANDLERS (which are used by the master control) did not have
an immediate effect in the CMC, only after (re-)loading the CMC config.

Change-Id: Ia3e84366f2b03bd2fcb522c2e10feaad0b01864d