Fail rename_and_wait_for_completion in case the rename_host job failed
So far the wait polling just ended without any follow-up check on the
success of the action.

The rename operation is a complex task which involves multiple actions
on the central and remote site. This is not a transaction which can be
rolled back. So a rename operation may actually be applied partially,
which is unfortunate. But this is the way it is right now.

We now want at least better error reporting about the failed job to
understand error situations easier.

CMK-20848

Change-Id: Ia443107dc6d0f6f20d414a62aecb88f3cc051432