Composition tests: Make test_automatic_host_removal more robust
Forces the Check_MK services to be in CRIT state

This makes the test independent from core scheduling and the whole check
cycle which shall not be tested here. Instead of relying on it, we
disable the core scheduling and set the services to CRIT state. This way
we can be sure that the conditions for the host removal are met.

Change-Id: I6df27ef5003d4e01ecae26923eb184e678fef120