17338 Stop early out on many select2 boxes, enable them asynchronously
Previously, when a HTML form page was rendered with a lot (>hundreds) of
searchable items, the augmentation process to make them searchable was
taking a very long time.

In Werk #15098, we addressed this by not making them searchable if too
many such items are present.

This was very noticeable when editing a notification rule with a
checkmk configuration containing many tag definitions for example.

With this werk, we restore the searchability if many searchable items
are present by letting the browser execute the augmentation process in
the background.

This process should not be noticeable by the user. No manual action is
required to make this work.

SUP-21833

Change-Id: Iac409030797c030bd03858eaf87294389e0053dd