Fix race condition in cmk-ui-job-scheduler
In the new UI job scheduler we have some jobs running in threads. Each
of those threads loads the UI configuration and uses thread local
proxies (active_config) to access the configuration, which is fine.

However, there are register_post_config_load_hook that are called once
the config was loaded. The update_config_based_user_attributes was
updating the global user_attribute_registry which lead to race
conditions between threads.

With this change we now get rid of the modification of the global state.
Instead we make the dependency on the config explicit. The call sites
now get a combined list of user attributes from the static global
registry and the config based user attributes.

Change-Id: Ieec3189cdc8d51c1040d9c622ea21755eda94565