Check all python packages with semgrep
We want to have semgrep coverage for all our code. So far semgrep only
covered the code outside of the packages.

Please note: We don't want to duplicate the rules, which are not only
enabling/disabling of predefined tests offered by the tool, but more
complex definitions of what is considered a problem in our code. There
is just one rule now, but that will grow over time and it does not make
sense to maintain them individually for every package.

Change-Id: I13ead71eaa0a67bb49fb22f2358d556611f87fb1