17339 Allow log-in with single session limit if previous session was invalidated
With the global configuration option "Limit login to single session at
a time", we allow limiting the active sessions per user.

This setting defines a time out until a new session can be opened.

Previously, this time out needed to be awaited before a new session
could be opened.

With this werk, we allow creation of a new session if the user has
logged out of a previous session.

Change-Id: I9ef98ffe20753d97a658655c86d4f171942e6141