ruff: Enable unused import (F401) rule
Unused imports add a performance overhead at runtime, and risk creating
import cycles and should be avoided. As we can see in this and the
following change, we accumulated quite a few unused imports in the short
time since the introduction of ruff. This shows pretty well that we need
to have this rule enabled.

Change-Id: Ib921d88242999a1fceba79b600a3d6e570575fad