.bazelrc: Limit CPU available in CI
In order to limit CPU usage peaks in our CI infrastructure, we limit the
amount of available CPUs per Bazel process to ~10.
This is used to ensure that most jobs are able to finish in time, while
we limit the amount of usable CPU resources to achieve good runtime even
with many parallel jobs running at the same time.

References CMK-20447

Change-Id: I8d59f8419d6a82911bbdc649d2578a33133660d2