17004 FIX Fix label filters for Quicksearch results
Using the "Quicksearch" sidebar element to look for hosts with a specific host label (e.g. with a search string "hl:label:key") links users to the "Host search" view, filtered by the specified label ("label:key").
Yet, adding more entries to the label filter on the "Host search" view was broken and resulted in an unfiltered view.

This werk fixes the above behavior, making label filters function as expected after use of the "Quicksearch" sidebar element with a label specifying string.

SUP-21749

Change-Id: I217439c1e370ccb6280ed96ab4230a18fce04f04