schemathesis tests: Fix undocumented Content-Type errors
In some cases, the webserver may return error messages in a
Content-Type that is not documented in the API schema.
This is not really an API issue an causing a lot of noise in
the tests. This change ignores the content type for problem+json
responses with the status code 400 and 404.

Change-Id: I5302a6f3641aaed01f1810b810a417b851e08b8f