utils(xml): swap out dicttoxml implementation
The following implementation is meant to replace the third-party
`dicttoxml` library. It's less featureful and more opinionated than the
original library, but should satisfy our current needs.

For instance, the new implementation hardcodes the following default
behavior from the former libary:
  - Root tag name is always set to "root".
  - Support for `cdata` and `ids` dropped.
  - The `type` attribute is always set for XML nodes.

Interestingly, the original implementation also allowed for the passing
of non-dictionary types to `dicttoxml`. This is no longer supported.
Instead, the input type is now constrained to a dictionary. This reduces
the complexity and scope of the function.

The following additional improvements were also made during the rewrite:
  - Reduction in code (474 -> 94 LOC).
  - Type annotations for maintainability and safety.
  - Deduplication of shared behavior (rendering of XML content).

CMK-20831

Change-Id: Ifba846f14e6fbcdd5a0efa6c8410e9af2d9abe05