Let Bazel manage taplo
This approach is in sync with how we add other dev tools to the project.
It is used by our test targets and can be invoked manually like this:

  bazel run //bazel/tools:taplo_bin -- [ARGS]

Since we now have a reliable automatic way to provide taplo, we can
make the so far optional taplo call mandatory.

Change-Id: I4a89f9240156f79b432dc4c2328e7ebc0f4ca4ba