tests.testlib: bugfix in `site.Site`; enables usage of `REUSE=1`.
Skip creation of (existing) automation user within an existing site
, when REUSE=1.

Details:

During a (system) test-run in the site creation stage,
a new automation user is being created.

When tests are re-run using `REUSE=1`, the site-creation is skipped
BUT the automation user creation is not. This led to a pytest run
failure, as the user already existed.

CMK-21035

Change-Id: I8352872e0766b32792999794916a14914e96a15b