Fix potential icon config related race condition
In the new UI job scheduler we have some jobs running in threads. Each
of those threads loads the UI configuration and uses thread local
proxies (active_config) to access the configuration, which is fine.

However, there are register_post_config_load_hook that are called once
the config was loaded. The update_icons_from_configuration was
updating the global icon_and_action_registry which lead to race
conditions between threads.

With this change we now get rid of the modification of the global state.
Instead, we make the dependency on the config explicit. The call sites
now get a combined list of user attributes from the static global
registry and the config based user attributes.

Change-Id: Ie6dbdbb957ab139247633f9c8e87f6401b33d5f8