ruleset matching: pass discovered labels
With this patch we get rid of the match object and
pass the service labels explicity to the respective
functions.

This also implies passsing the _discovered_ service
labels explicitly.

In the past, we had to read the autochecks and compute all
their service descriptions to find the correct discovered
labels -- an information that was previously already known.

Change-Id: I8aa018f4820d0038116058e9c38f17f0bae55982