gui_e2e tests: improve assertion/expect messages
(main reason was to get comfortable with playwright and pytest)

about the rational for this change:
Assertions/expects without explicit message are often hard to decypher for developers, because they mostly focus on locators and too seldomly are based on the meaning for the user.
Thus these assertions were made readable and understandable by developers.

CMK-20723

Change-Id: Id7c926cfcc805d2a8ff3dac7807432f71630df6a