17483 FIX missing ruleset from Azure Databases
The ability to set levels for the average deadlock count was simply
missing. In this change, we keep with the status-quo of not setting any
default values for the thresholds. Although none of the other Azure
Database plugins make use of this parameter, I've added it to all of
their default parameters so as to not deviate from the existing pattern.

SUP-21803

Change-Id: I7426f0b2de74fc4a3dea020a9f27e0ff924f18c6