17300 Windows agent do not use timeout on fail for section CpuLoad and DotNetClr
Previously, after a WMI error occurred during the generation of
specific WMI sections, particularly the CpuLoad and DotNetClr sections,
the agent would stop creating these sections for a period of one
hour. This was done to reduce noise in the log file. However,
this approach led to situations where a random WMI failure could
prevent data collection for an hour.

Since this release the delay for the CpuLoad and DotNetclr sections is
eliminated thus making monitoring more smooth.

Change-Id: I12f92a5a8d7a517e91f75d7c07f523917df33aae