gcp_cost: only query current month
Previously also the costs of the previous month were queried, and shown
in the details. With this change it is no longer shown in details,
please refer to the newly introduced metric if previous cost development
is important for you.

This should reduce the cost of querying the cost significantly.

Change-Id: I6e92b3e305f8236a338956f2c2db1120dc90bd5f