test_plugin_piggyback: Reduce test's flakyness
We here remove any assertion over pending changes during the activation
of the DCD connector. This is because the creation of the DCD connector
initialize several new changes that might need quite some time to be
activated.
The changes initialiazed by the DCD connector are anyway handled within
the test. Thus, there is no need to be strict about them beforehand.

Change-Id: Ia72e5ae9ae715fd6f12935e8bbe2ba9b10faf7e9