15352 gcp_cost: several changes
Choosing predictive levels for gcp_cost resulted in a crash of the check. This
option was removed. If you have a rule using "Predictive levels" for gcp_cost it
will be updated to be "No Levels".

Previous to this change it was possible that not all projects were available for
monitoring.

The old query did not include credits based costs.

Each check now creates a metric of the current costs.

Previously the check showed the accumulated costs for the current month, and the
total costs of the previous month. Each query costs money based on how many rows
are accessed, so showing the costs of the whole previous month is costing money,
so we decided to remove this output. As we added the metric, it is now possible
to visually compare the current month's costs with the previous month's cost.

Change-Id: I999acde7d8b81b906304dba51df78361d62942b4