Move automation helper activation switch to "Site management" settings group
Previously, this setting was localted in the "Developer tools" group. The
settings in this group come with a warning that any changes to them will render
the site unsupported. This is not the case for the automation helper. Customers
are allowed to deactivate them, especially since they are still brand-new.

Change-Id: I90cd41d7136f1a9bc2a2cccd181cc32cd8fb37cd
(cherry picked from commit 5b36abbb2474f39803d2006ff39c937d670960d5)