mk-sql: refactor config.rs  to use include_str! in unit test
Since file paths for test data differ for run time test data between Bazel and
Cargo it is easier to load all data at compile time. If we NEED to test file
handling then we can do this with tmp files if we need to keep Cargo and Bazel.

Since we are not using files in these tests, added a convience wrapper
"load_str" which is identical in handling to "load_file".

Also since this code needs to use an internal API call we make function
"pub create"

CMK-19040

Change-Id: I0ddbf2fc38dc33506c0038f739881e337279c7b3