17369 FIX Use correct ports for netapp's failover in case of "home_port_only"
You are affected by this werk if:
* you're using the netapp ontap special agent
* you're using the failover policiy `home_node_only`
* you have used the deprecated special agent via webapi and set custom failover groups.

Before this werk, the `home_node_only` policy would have taken all ports of the home node into consideration.
As per netapp's documentation, failover groups do not exists anymore in the rest API:
> The common practice will be to use default failover groups, which are all the ports of a broadcast domain. Custom failover groups are not presented in the REST API.

The consequence is, that the ports of the broadcast domains are now considered and only those which are on the home node.

Change-Id: I7ea093920ccab44502a64901e0a52db423b44363
JIRA-Ref: SUP-21666