Prevent automation related crashes in openapi tests
Some openapi unit tests trigger automation calls, which in turn trigger
crash reports because the `check_mk` command is not available in the unit
test context.

The crash reports were unfortunately not reported by the tests. These
issues were only visibile with CLI logging enabled. This will have to
be cleaned up by disabling the crash reporting in the WSGI App in a
follow-up.

Change-Id: I9179c0288230a88545c91aca86ac635178111c84