Improve ui-job-scheduler startup
This change does not actually improve the startup time since we still
load the whole UI code, but moves the loading the UI code from pre to
post daemonize. This makes 'omd start' faster at the price that
potential initialization errors will be written to the web.log instead
of the console.

CMK-21190

Change-Id: Ida0071c16fc3cad4e093c08bd70c2732df611cbc