Keep ccc libraries separated
This package is a collection of (mostly) unrelated libraries.  Let's
reflect this in Bazel to simplify and clean the dependency tree.

This package will also certainly grow with time and pull an increasing
amount of dependencies.  Therefore, keeping the components minimal
should have a positive impact in the long run.  This will further
simplify our work when we decide to refactor (or remove) some of these
cross-cutting concerns.

Change-Id: I1265ca34e565f0c06fab8030eaee2f0c32fb5658