check-cert: Test "truststore::system()" does not fail.
Testing that "truststore::system()" does not fail. is valid, testing the content
assumes the value of the content, which will be different on a hermetic build
from a developers computer.

Bazel attempts to do hermetic builds, cargo not, we want both to prodcue the
same result.

By hermetic builds we mean that they are insensitive to the libraries and other
software installed on the build machine.

CMK-19040

Change-Id: Ib2406ef645602b2e0fbc028ec61ab8fc59f780a9