Automation helper: New algorithm for reload decision
Before, after a change was detected, we waited 5 seconds and then reloaded. This
algorithm leads to constans reloads every 5 seconds in very dynamic
environments.

To mitigate this, we introduce a new algorithm: After a change is detected, we
wait 5 seconds. If no further changes occured within this cooldown period, we
reload. Otherwise, we compute the time difference between the last detected
change and the change that we initially detected and wait for this additional
time difference. Then we check again for new changes that came in in the
meantime and apply the same procedure until we encounter a 5 seconds window
without any changes. Then we reload.

|-------------|--------------------|-------------|
t1            t2
change        another change       5s cooldown   5s + t2 - t1
detected      detected             over          now we reload
                                   no reload

CMK-21331

Change-Id: I345e79be364990d23df277fae9dcd662a3de2ba5