cmk-agent-ctl: Use a feature to define how to include common code for tests.
Cargo allows a shared code in the rust integration tests.
Bazel does not support shared code in the rust integration tests.

To work around Bazel we make a shared library just for integration tests. This
then has to be included differently between Cargo and Bazel.

We add the 'build_system_bazel' feature to the bazel build to allow Cargo and
Bazel to both work.

CMK-19040

Change-Id: Icbb70efb138993b0f9ba2a7a9f449f188f63c4ad