mk-sql: use feature for integration tests
Cargo allows a shared code in the rust integration tests.
Bazel does not support shared code in the rust integration tests.

To work around Bazel we make a shared library just for integration tests. This
then has to be included differently between Cargo and Bazel.

We add the 'build_system_bazel' feature to the bazel build to allow Cargo and
Bazel to both work.

CMK-19040

Change-Id: I2b0ae5c1695e020a888642e5db45be397b846ae4