mk-sql: export TempDir from shared test code
Cargo allows a shared code in the rust integration tests.
Bazel does not support shared code in the rust integration tests.

To work around Bazel we make a shared library just for integration tests.

When the tests are using a shared library the version of TempDir may potentially
not match. For this reason we must export the TempDir type from the shared
library and use this Type consistently.

When it is all the same code / module this is not needed. But we want to use
Bazel and Cargo transparently.

CMK-19040

Change-Id: I22ab9a62b62ada3a6e5cb004d480f2c8ac71faff