Improve ui-job-scheduler logging
The appplication logs now go to `var/log/ui-job-scheduler/ui-job-scheduler.log`
instead of `var/log/web.log`. This is to avoid confusion with the logs emmited
by the UI code run in the context of the apache.

Additionally the thread name is now included in the log messages to make
it easier to track the execution of the individual threads.

Change-Id: Ie5649ba680952f8f27498bea517a497225322d55