Prepare ui-job-scheduler to receive commands
So far the process was just following the hard coded schedule of the UI
jobs registered to the application and caring for their execution.

There is another type of job however, which was executed directly from
the UI application, the background jobs. We have the need to start and
observe longer running tasks from the UI application.

These background jobs are executed using multiprocessing.Process. We
used the fork mode before 2.3 and moved it to spawn in 2.4, because we
experienced issues with leaking resources, like locks or other open file
handles. Now, the spawn mode is better when it comes to resource
management, but comes with a huge initialization overhead compared to
fork, which we want to avoid.

Now the idea is to extend the capabilities of the job scheduler, which
has all the UI application context already initialized and let the
background jobs run as threads in the same process. This way we can
avoid the overhead of spawning new processes and still have a clean
decoupling from the apache process.

For the interface between the UI and the job scheduler, we will reuse
the same components as we used for the automation helper.

Change-Id: I72309ddde88984ebf4ead2a8ff14cf3e1e64a9ed