Call super constructor first
In marshmallow 3.24.0, Field.parent is not a class
attribute anymore, but an instance attribute, see
https://github.com/marshmallow-code/marshmallow/pull/2723.

In any case, always calling super().__init__(...) and as
the first statement of a constructor is a very good idea.
Otherwise your class hierarchies involving the class in
question can and *will* break sooner or later due to the
way the MRO works.

Change-Id: I414be7062822f984444d2508061e9c2885860d88