Fix initial agent baking and REST API spec computation
The initial agent baking was started before the site was
started, which does not work anymore with the background jobs
now being dependent on the UI job scheduler process.

We now just place a signal file to trigger the agent baking
once the job scheduler is ready.

The REST API spec uses the same mechanism and is not triggered
during apache startup anymore.

Change-Id: I0a1d2c38f171a3d501708de73766291a4264cd86