Execute background jobs as threads
The so far used subprocess spawn method in 2.4 comes with too much
overhead. This change introduces a new way to execute background jobs as
threads. This will only work properly if the threads are executed in the
context of a continously running process, like the ui-job-scheduler. The
following changes will move the background job execution to this
process.

This change enables us to re-enable the background job tests (see
CMK-18161).

Change-Id: I550fa4e184b80064831cdfb61949c027860a037d