17802 FIX liebert: Change metric names to use correct general graph definition
This Werk changes the names used for the metrics in the plugin, so that they use the same graphing rules as other metrics that output fan percentages.

This change is incompatible as it will create new graphs and not continue the old graphs.

Thanks to indigoparadox for the change.

CMK-21471

Closes #783

Change-Id: I784284e429f494c919815ab4f7e8e7d72b14f474