ui-job-scheduler: make race condition less obvious
There is a 5+second window, where the following happens:

```
$ omd start ui-job-scheduler
Starting ui-job-scheduler...OK
$ omd status ui-job-scheduler
ui-job-scheduler:   stopped
```

That is because the child process takes 5+ second until it creates the
PID file, `omd status ui-job-scheduler` uses to determine the status.

Really, the parent should wait for the file, but `daemonize` does not
appear to support that functionality.

CMK-21453

Change-Id: Ie10decfb107b887a99a36bc13c05596e0625a6ea