17043 FIX extreme_vsp_switches_temperature: Fix 'Cannot render negative timespan' crash
This affects user that have the parameter 'Levels on the time left until a critical temperature (upper or lower) is reached' configured.

Currently, if the temperature was above the set upper critical level or below the set lower level, the check crashed with an error similar to:

```
File "/omd/sites/vls/lib/python3.12/site-packages/cmk/agent_based/v1/render.py", line 76, in _gen_timespan_chunks
    raise ValueError("Cannot render negative timespan")
```

This has now beed fixed. With the new behaviour if the critical levels have already been reached, it will be represented in the summary.

Background: If the critical levels were already reached, it caused the
seconds_left variable to be negative, which caused the crash.

SUP-21709

Change-Id: Ib0caf96f26a687b5196275b50a2570bfe3531d55