17577 FIX virtual_host_trees: match on tag group and tag group tag id
Previously when a Virtual host tree was configured, the
tree would be built matching on the tag id.  However,
tag ids can be the same across tag groups which was not
taken into account. With this werk, we now match on
tag group and the tag group tag id.

SUP-18766

Change-Id: I4eb43dffc18b6cde09fd74dd5f8349d271b229f5