Make scrollbar style active by default across vue
The previous implementation relied on all possible occurences of
scrollbars to be wrapped in <CmkScrollContainer>, this led to many
missed instances (any resizable textbox for example).

To not interfere with our legacy UI, we add a global css rule to define
our default scrollbar under the .cmk-vue-app class.

The CmkScrollContainer can still be used to create the "outer" scroll
bar style etc.

Unfortunately, due to the slidein creating elements outside of our "app"
div, we need to supply the class here manually.

CMK-21301

Change-Id: I9dab415800a389d93091c7b1edb3f0ba7d7e326f