Base-align all FormApp labels
This is a no-op for 99% of all labels, since most labels
are rendered next to an input field of some sorts and by default are
vertical-aligned: baseline.

For the cases where we have a label followed by a line break, this will
give them the same padding that they would have when they would be
placed next to an input element.

A visualization of the problem can be seen in the CascadingSingleChoice
example screenshot in the ticket.

CMK-21301

Change-Id: I70bd1c7ffa1ed1b902d46800044880a47078daac