watolib(automations): handle exit codes
Prior to this change, we always raised an `MKAutomationException` when
an automation call exited with a non-zero code. This was problematic
because each time a call failed, it would generate a crash report.

With this change, we now distinguish between known/general errors which
we can expect in an automation call vs unknown errors which we haven't
accounted for. The former raises the `MKUserError` which will still
inform the user about the issue, but won't generate a crash report.
Whereas, the latter continues to crash on failure as before.

CMK-21313

Change-Id: I551e868a190235daf79cab673a86e9ed45dcb7fc