Fix inputs inheriting margin from legacy CSS
It's an ubiquitous pattern to have <table class="nform"> in our
codebase, this triggers a css rule in our non-vue CSS file which
adds padding to all inputs. Our vue apps are frequently also
rendered in such a table.

Since we cannot touch that rule else we'd risk changing the non-Vue UI,
we compromise by overwriting the rule in our global Vue CSS.

The effect of this was very noticeable in Tuple / Levels FormSpec.

This also allows removal of the floating text class.

CMK-20645

Change-Id: Ib1bac5a641032f6a81b636d4156992d2a54922fe