17578 FIX notifications: 'notification_bulking' not a required field
For most notification plugins, the notification_bulking field didn't make
sense and was previously made a non-required field.  However, if you
tried to create a notification rule via the rest-api without this field,
it would generate a crash report.  This werk addresses this issue by first
checking if the field is present before setting it's value.

Change-Id: I098e9929be0a2f38715b18c0c253dfcd3aa48c0c