watolib(automations): return message not exc
The `local_automation_failure` function proved to be too opinionated by
always returning `MKAutomationException`. We want more flexibility so as
to handle different failure codes with different exceptions. Therefore,
this helper now only builds an error message based on its input params,
and it's now the job of its caller to select an exception to raise.

CMK-21313

Change-Id: Ia07d26390114f716d94e70a02292fa1b144459ba