automation_helper: don't write to tty
We cache whether we are attachted to a tty at import time of
`cmk.utils.tty`. For the automation-helper this means that all
automations will print colors, since at import time `sys.stdout.istty()`
is `True`. In particular, this change fixes the symbols in the
diagnostic dump.

Sven successfully predicted this bug, see the TODO at the top of
`cmk/utils/tty.py`.

Change-Id: I382f388c43a7e7d3868a9c7018ec243b0669ce2a