Job scheduler: Improve logging
Previously the logging was initialized with INFO level and later
overwritten by set_log_level(), which is executed while initializing
the gui context for the individual jobs. Actually, this is a side effect
of that context manager which we should to clean up. The current
situation can be really confusing. But this is something for another
day.

To reduce the confusion, we now initialize the logging with the UI log
levels and use a new dedicated UI logger for the job scheduler.

Additionally, temporarily set the log level to DEBUG for the system
level tests.

Change-Id: I03a1fd97ff66b1082daf8b0984e4c273de786a3b