Remove gunicorn from cmk-ui-job-scheduler
Instead of using a combination of gunicorn + uvicorn, we simplify
the application server setup by using uvicorn only. This reduces the
memory footprint and the complexity of the setup since we only use
a single process now.

We may have to adjust the number of workers in the future to ensure
that the application server can handle the incoming requests or change
the endpoint implementation to not block the asynchronous event loop.
But for now, we will start with a single worker, which is equal to the
gunicorn setup we had before.

Change-Id: I3bad5bd584095d99cd043153d05f9764233d8e9f