Bring back error handling of background job start
In one of the recent refactorings the error handling of the start method
was changed and the calls sites were not updated, which lead to errors
not being reported to the user. This change tries to fix this, but the
specific ways the errors are handled might need to be adjusted in a few
places.

Change-Id: I2b872239834c3d7ce48ed101678a1432d5fb18b1