remove hover effect for textarea fields
the rule seems to add a hover effect for all input fields, but this did
not work for text input fields and select elements, as it was
overwritten by other rules. it actually worked for textarea and
checkboxes. indicating an clickable element for checkboxes seems
reasonable, other selectors were removed.

CMK-21237

Change-Id: Ia2fddaf5fc86506cfbef0a9587a4c95152b6abae