Fix get-configuration automation not returning defaults
While porting the get-configuration automation to run on the automation
helper, the loading of the non-confd configuration was removed. This
however broke the only callsite of this automation which was to get the
factory settings of our configuration.

We restore the old behavior of the automation by forcing it to run
outside the automation helper.

To still have some performance improvements for the global settings
page, we cache the output of this automation call in the apache process.

Slightly risky :)

co-authored-by: benjamin.knapp@checkmk.com

CMK-21350

Change-Id: Ia0d8144f77ea53ef9bddcbd151dd15f41bcbe916