`testlib`: Make global settings update more flexible
Not all global settings go to the same file. Therefore, the target file is now
an argument to the respective functions.

This change also corrects the file to which the agent bakery logging settings
are written when setting up for the composition tests. Before, these settings
were written to the wrong file.

Change-Id: I45b8dcd698b84c56ec137f735c636f0adb5851bd