Regression: ship (again) pyc files
This was basically broken when the build process of python3-modules was fully integrated into bazel.
We now use the python interpreter from bazel to build those files and not anymore the one we ship.
At the moment we believe this is OK as:
* we're using the same python version
* we're using the checked-hash invalidation mode

Change-Id: Ife235a857a482c112c3cf142d4b9d5b7a4b6cbb2
JIRA-Ref: CMK-21242
(cherry picked from commit 32999ed1ed25259e07f1ad011dbe90886def3e82)